Add more code cleanup from intellij code analysis

+++ src/main/java/com/marklogic/spring/batch/core/MarkLogicJobInstance.java
@@ -74,9 +74,9 @@ public class MarkLogicJobInstance {
 				jobExecutionIndex = jobExecutions.indexOf(je);
 			} 
 		}
-		if (jobExecutionIndex.intValue() >= 0) {
+		if (jobExecutionIndex >= 0) {
 			jobExecutions.remove(jobExecutions.remove(jobExecutionIndex.intValue()));
-			jobExecutions.add(jobExecutionIndex.intValue(), jobExecution);
+			jobExecutions.add(jobExecutionIndex, jobExecution);
 		} else {
 			throw new NoSuchObjectException("JobExecution " + jobExecution.getId() + " does not exist");
 		}
